Rename footer and header JS files
This was done because I'll put all the JS for each component in the
same file, so that component.js will (at least) two functions: the one
that handles the component's logic and the one that handles the
component's DOM logic. This will keep import statements a bit cleaner
in the index.js file.